Tickets

+++ app/(pages)/projects/index.ts
@@ -26,10 +26,29 @@ export async function postProject({
 	return data;
 }
 
-export async function deleteProject({ projectId }: { projectId: string }) {
+export async function deleteProject({ projectId }: { projectId: string | undefined }) {
+	if (!projectId) return;
 	const response = await fetch(`/api/projects/?projectId=${projectId}`, {
 		method: "DELETE",
 	});
+	const data = await response.json();
+	return data;
+}
+
+export async function assignProject({ projectID, userID }: { projectID: string; userID: string }) {
+	const response = await fetch("/api/project_assignment", {
+		method: "POST",
+		headers: {
+			"Content-Type": "application/json",
+		},
+		body: JSON.stringify({ projectID, userID }),
+	});
+	const data = await response.json();
+	return data;
+}
+
+export async function fetchRole({ userID, assignedID }: { userID: string; assignedID: string }) {
+	const response = await fetch(`/api/user_role?userID=${userID}&assignedID=${assignedID}`);
 	const data = await response.json();
 	return data;
 }
